==> report/01-Data.Rmd <==
---
title: "Assessment 0 - 01: Data and Introduction"
output: html_notebook
---

## The brief

The brief was to:

* generate a sequence of visualisations of *conn_sample.log* which was presented in class in Week 1. This is the Bro log data from [Secrepo](http://www.secrepo.com/Datasets%20Description/HTML_Bro_log_1/conn.html).
* Make a polished figure with a caption that could be used to describe the dataset in a publication.

## The team

This is a two person team, *Rachel* and *Peter*. Rachel is an expert in R and Peter has some experience in Python. We have agreed to split the work by coding language. We both intend to contribute 50% to the project.

(Comment: You will see from the reports that the R content is better than the python. Perhaps Peter is worried that his python skills will let the side down, he intends to try to make up for this by contributing mathematical skills and some python is better than none. If the attempt is genuine, the evidence of effort and some contribution is present in the individual content directories, and the group agree that an equal contribution has been made, this will be respected.)

## Library requirements

This installs missing packages required for this script only. Each script handles it's own requirements, using a simple check-and-install system. 

(Comment: Note that this does **not** guarantee that the correct version of packages is installed; for that, a `requirements.txt` file can still be created and the installation automated; see https://stackoverflow.com/questions/54534153/install-r-packages-from-requirements-txt-file. There is a library called [packrat](https://rstudio.github.io/packrat/) which is supposed to replicate python's virtual environment functionality, though it may be more cumbersome than required for our needs.

NB Your code should always run on Unix-like systems, though some leniancy will be granted.)

We need `fs` to be able to handle directories in a cross-platform way. This makes the whole analysis much more repeatable and is good data science practice.

```{r}
if (!require("fs")) install.packages("fs")
library("fs")
```


## The data

We were assigned a dataset in the brief: **conn_sample.log** data, the dataset presented in Week 1 Lectures.

To obtain this dataset in a convenient format for both analyses, we will download it and process it into a standard form (containing headers).

We place raw data in the `data` folder of our root. (Note: we are in the /report directory now). So our file system will look like this:

* /data
  * /data/raw
    * /data/raw/conn_sample.log
    * /data/raw/conn_sample.names
  * /data/processed
    * /data/raw/conn_sample.tab

### Get the data

First create a directory for it to go in:
```{r}
rawdatadir=path_wd("..","data","raw")
if(!dir.exists(rawdatadir)) dir.create(rawdatadir,recursive = TRUE)
```

Download it
```{r}
download.file("https://github.com/dsbristol/dst/raw/master/data/conn_sample.log",path_wd("..","data","raw","conn_sample.log"))
```

### Get the names for the data

NB We had to scrape this from the website which is not very elegent, but works.

```{r}
cnames=c('ts','uid','id.orig_h','id.orig_p',
    'id.resp_h','id.resp_p','proto','service','duration',
    'orig_bytes','resp_bytes','conn_state','local_orig',
    'missed_bytes','history','orig_pkts','orig_ip_bytes',
    'resp_pkts','resp_ip_bytes','tunnel_parents')
write.table(cnames,
            file=path_wd("..","data","raw","conn_sample.names"),
            row.names=FALSE,
            col.names=FALSE,
            quote=FALSE
            )
```

### Create the processed data output.

```{r}
processeddatadir=path_wd("..","data","processed")
if(!dir.exists(processeddatadir)) dir.create(processeddatadir,recursive = TRUE)

conndata=read.table(path_wd("..","data","raw","conn_sample.log"),
                    as.is=T)
colnames(conndata)=cnames
write.table(conndata,
            path_wd("..","data","processed","conn_sample.tab"),
            quote=FALSE,row.names=FALSE)
```

### Summary:

We have now created our data for further processing, in a simple headed tabular format (which is space separated). This can be easily read into either R or Python.
