==> report/04-Wrapup.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

## Wrapup 

We have been able to explore network visualisations, t-sne, balloonplots and other complex visualisations on our data. Throughout the challange has been in figuring out what parts of our data can be given to the various visualisation techniques to get it working.

For creating our figure, we selected the "ports and services" balloonplot as it was most complementary to the figure from the python notebook that was selected. Here it is finally:

```{r}
library("fs") # for cross-platform directories
portservicedata=readRDS(file=path_wd("..","data","processed","portservice2.RDS"))
par(las=2)
balloonplot((log(1+portservicedata)), main ="ports and services", xlab ="", ylab="",
            label = FALSE, show.margins = FALSE,text.size=0.5)
```
## Conclusions

These data are strange and we have explored many tools for understanding them better. We updated python code to work on our data, and found several R resources which we documented throughout.
